First solution to function inlining
Reducing type size and avoiding the call with a lambda.
Added debug/release to build.sh

+++ labs/core_bound/function_inlining_1/solution.h
@@ -7,8 +7,8 @@
 constexpr size_t N = 10000;
 
 struct S {
-  uint32_t key1;
-  uint32_t key2;
+  uint16_t key1;
+  uint16_t key2;
 };
 
 void init(std::array<S, N> &arr);
